squash! cxl/mem: Read dynamic capacity configuration from the device
Devices which optionally support Dynamic Capacity (DC) are configured
via mailbox commands.  CXL r4.0 section 9.13.3 requires the host to issue
the Get DC Configuration command in order to properly configure DCDs.
Without the Get DC Configuration command DCD can't be supported.

Implement the DC mailbox commands as specified in CXL 4.0 section
8.2.10.9.9 (opcodes 48XXh) to read and store the DCD configuration
information.  Disable DCD if an invalid configuration is found.

Linux has no support for more than one dynamic capacity partition.  Read
and validate all the partitions but configure only the first partition
as 'dynamic ram A'.  Additional partitions can be added in the future if
such a device ever materializes.  Additionally it is anticipated that no
skips will be present from the end of the pmem partition.  Check for and
disallow this configuration as well.

Linux has no use for the trailing fields of the Get Dynamic Capacity
Configuration Output Payload (Total number of supported extents, number
of available extents, total number of supported tags, and number of
available tags).  Avoid defining those fields to use the more useful
dynamic C array.

Based on an original patch by Navneet Singh.


---
Changes:
1. Move partition alignment check after is_power_of_2() check on
   blk_size, as IS_ALIGNED(partition start, blk_size) expects blk_size
   to be a power of 2 in cxl_dc_check()

2. cxl_get_dc_config(): verify mbox_cmd.size_out against
   dc_resp->partitions_returned

3. cxl_dev_dc_identify(): originally calculated size of dc_resp using
   struct cxl_dc_partition_info, but dc_resp->partition[] is of type
   struct cxl_dc_partition. Fix size calculation.

4. fix do/while loop in cxl_dev_dc_identify to protect against returning
   0 partitions infinitely

5. cxl_configure_dcd(): originally checked for gap between PMEM and DC
   partition by calculating if a gap exists:
   	if ([start of dc part] - [end of pmem part])
   Replace with: if ([start of dc part] != [end of pmem part])

6. Change struct cxl_dc_partition_info to use u64 instead of size_t
   fields

7. Original commit message referenced CXL r3.2. Bump to r4.0.
   Verified section numbers remain the same

+++ drivers/cxl/cxlmem.h
@@ -902,8 +902,8 @@ int cxl_internal_send_cmd(struct cxl_mailbox *cxl_mbox,
 int cxl_dev_state_identify(struct cxl_memdev_state *mds);
 
 struct cxl_dc_partition_info {
-	size_t start;
-	size_t size;
+	u64 start;
+	u64 size;
 	u8 handle;
 };
